[#258] Refactor test package
