+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/data/remote/models/responses/user_response.dart
