+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/data/repositories/credential_repository_impl.dart
@@ -1,4 +1,4 @@
-import 'package:{{project_name.snakeCase()}}/data/datasources/remote/api_service.dart';
+import 'package:{{project_name.snakeCase()}}/data/remote/datasources/api_service.dart';
 import 'package:{{project_name.snakeCase()}}/domain/exceptions/network_exceptions.dart';
 import 'package:{{project_name.snakeCase()}}/domain/models/user.dart';
 import 'package:{{project_name.snakeCase()}}/domain/repositories/credential_repository.dart';
