+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/di/module/network_module.dart
@@ -1,4 +1,4 @@
-import 'package:{{project_name.snakeCase()}}/data/datasources/remote/api_service.dart';
+import 'package:{{project_name.snakeCase()}}/data/remote/datasources/api_service.dart';
 import 'package:{{project_name.snakeCase()}}/env.dart';
 import 'package:{{project_name.snakeCase()}}/di/provider/dio_provider.dart';
 import 'package:injectable/injectable.dart';
