+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/app/screens/home/home_view_model_test.dart
@@ -6,7 +6,7 @@ import 'package:{{project_name.snakeCase()}}/domain/usecases/base/base_use_case.
 import 'package:{{project_name.snakeCase()}}/app/screens/home/home_screen.dart';
 
 import '../mocks/generate_mocks.mocks.dart';
-import '../mocks/response/user_response_mocks.dart';
+import '../mocks/data/remote/models/responses/user_response_mocks.dart';
 
 void main() {
   group("HomeViewModelTest", () {
