+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/data/repositories/credential_repository_test.dart
@@ -1,7 +1,7 @@
 import 'package:{{project_name.snakeCase()}}/domain/exceptions/network_exceptions.dart';
 import 'package:{{project_name.snakeCase()}}/domain/repositories/credential_repository.dart';
 import 'package:{{project_name.snakeCase()}}/data/repositories/credential_repository_impl.dart';
-import 'package:{{project_name.snakeCase()}}/data/datasources/remote/response/user_response.dart';
+import 'package:{{project_name.snakeCase()}}/data/remote/models/responses/user_response.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
