+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/mocks/data/remote/models/responses/user_response_mocks.dart
@@ -1,4 +1,4 @@
-import 'package:{{project_name.snakeCase()}}/data/datasources/remote/response/user_response.dart';
+import 'package:{{project_name.snakeCase()}}/data/remote/models/responses/user_response.dart';
 
 class UserResponseMocks {
   static UserResponse mock() {
