+++ sample/lib/di/module/network_module.dart
@@ -1,4 +1,4 @@
-import 'package:sample/api/api_service.dart';
+import 'package:sample/data/remote/datasources/api_service.dart';
 import 'package:sample/env.dart';
 import 'package:sample/di/provider/dio_provider.dart';
 import 'package:injectable/injectable.dart';
