+++ sample/lib/model/user.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:sample/api/response/user_response.dart';
+import 'package:sample/data/remote/models/responses/user_response.dart';
 
 class User extends Equatable {
   final String email;
