+++ sample/test/api/repository/credential_repository_test.dart
@@ -1,6 +1,6 @@
 import 'package:sample/domain/exceptions/network_exceptions.dart';
 import 'package:sample/api/repository/credential_repository.dart';
-import 'package:sample/api/response/user_response.dart';
+import 'package:sample/data/remote/models/responses/user_response.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
