+++ sample/test/viewmodel/home_view_model_test.dart
@@ -7,7 +7,7 @@ import 'package:sample/main.dart';
 import 'package:sample/model/user.dart';
 
 import '../mocks/generate_mocks.mocks.dart';
-import '../mocks/response/user_response_mocks.dart';
+import '../mocks/data/remote/models/responses/user_response_mocks.dart';
 
 void main() {
   group("HomeViewModelTest", () {
